tests: move port distributor out of neon_fixtures.py
