+++ test_runner/regress/test_metric_collection.py
@@ -13,10 +13,10 @@ from fixtures.neon_fixtures import (
     PSQL,
     NeonEnvBuilder,
     NeonProxy,
-    PortDistributor,
     VanillaPostgres,
     wait_for_last_flush_lsn,
 )
+from fixtures.port_distributor import PortDistributor
 from fixtures.remote_storage import RemoteStorageKind
 from fixtures.types import TenantId, TimelineId
 from fixtures.utils import query_scalar
